+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This version adds a dependency on Swift.
 
 ### Breaking Changes
 
-- Make public APIs Swift friendly (#)
+- Make public APIs Swift friendly (#2416)
     - Renamed `SentrySDK.addBreadcrumb(crumb:)` to `SentrySDK.addBreadcrumb(_ crumb:)`
     - Renamed `SentryClient.add(_ crumb:)` to `SentryClient.addBreadcrumb(_ crumb:)`
     - Renamed `SentryClient.add(_ attachment:)` to `SentryClient.addAttachment(_ attachment:)`