+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Breaking changes
 
-- Renamed `SentrySession.errors` to `SentrySession.hasErrors`, and changed to be a boolean (#2410)
-- Renamed `- [SentrySession incrementErrors]` to `- [SentrySession encounteredError]`
+- Rename `SentrySession.errors` to `SentrySession.hasErrors`, and changed to be a boolean (#2410)
+- Rename `- [SentrySession incrementErrors]` to `- [SentrySession encounteredError]`
 
 ### Fixes